@@ -3,7 +3,7 @@ name = "youtube-download-helper"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.13"
 
 dependencies = [
     "yt-dlp>=2024.12.6",
